% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{call_completion_stats_by_provider}
\alias{call_completion_stats_by_provider}
\title{Gets call completion stats by provider}
\usage{
call_completion_stats_by_provider(lead_provider, incidents, dates, pb)
}
\arguments{
\item{lead_provider}{Lead provider}

\item{incidents}{incidents tibble}

\item{dates}{Vector of dates}

\item{pb}{progress bar}
}
\value{
Table of stats by day
}
\description{
Gets call completion stats by provider
}
