% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_alt.R
\name{call_completion_stats_by_provider_alt}
\alias{call_completion_stats_by_provider_alt}
\title{Gets call completion stats by provider}
\usage{
call_completion_stats_by_provider_alt(incidents_by_lead, dates, pb)
}
\arguments{
\item{dates}{Vector of dates}

\item{pb}{progress bar}

\item{lead_provider}{Lead provider}

\item{incidents}{incidents tibble}
}
\value{
Table of stats by day
}
\description{
Gets call completion stats by provider
}
