% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartCompletion-package.R
\docType{package}
\name{ChartCompletion-package}
\alias{ChartCompletion}
\alias{ChartCompletion-package}
\title{ChartCompletion: Visualizes chart completion rates over time}
\description{
Visualizes chart completion rates over time to better understand the effectiveness of interventions and improvements.
}
\keyword{internal}
