% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_chart_completion_stats}
\alias{get_chart_completion_stats}
\title{Get Chart Completion stats by day}
\usage{
get_chart_completion_stats(incidents)
}
\arguments{
\item{incidents}{incidents range to compute}
}
\value{
tibble
}
\description{
Returns tibble with an observation for each day in the incidents data.
All data references a 30-day rolling window.
}
\details{
Variables:
\itemize{
\item Open24Hr: Number of charts within past thirty days open for more than 24 hours.
\item Open4Day: Number of charts open for more than 4 days.
\item Open8Day: Number of charts open for more than 8 days.
\item AllIncomplete: Total number of incomplete charts
\item Incomplete30D: Total number of incomplete charts
\item total_calls: Total charts within past thirty days
\item incomplete30d_prop: Proportion of all calls within past thirty days taht are open
\item Over24Hr_Prop: Proportion of all calls within past thirty days open for more than 24 hours
\item Over4D_prop: Proportion of all calls within past thirty days open for more than 4 days
\item Over8D_prop: Proportion of all calls within past thirty days open for more than 8 days
}
}
