% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_alt.R
\name{get_open_chart_stats_alt}
\alias{get_open_chart_stats_alt}
\title{Gets Updated Open Chart Stats}
\usage{
get_open_chart_stats_alt()
}
\value{
tibble of stats
}
\description{
Returns tibble with an observation for each day in date range 2021-09-01
through tthe current date. Saves as data.
}
\details{
All data references a 30-day rolling window.

Variables:
\itemize{
\item Open24Hr: Number of charts within past thirty days open for more than 24 hours.
\item Open4Day: Number of charts open for more than 4 days.
\item Open8Day: Number of charts open for more than 8 days.
\item AllIncomplete: Total number of incomplete charts
\item Incomplete30D: Total number of incomplete charts
\item total_calls: Total charts within past thirty days
\item incomplete30d_prop: Proportion of all calls within past thirty days taht are open
\item Over24Hr_Prop: Proportion of all calls within past thirty days open for more than 24 hours
\item Over4D_prop: Proportion of all calls within past thirty days open for more than 4 days
\item Over8D_prop: Proportion of all calls within past thirty days open for more than 8 days#'
}
}
