% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_chart_by_provider.R
\name{graph_chart_by_provider}
\alias{graph_chart_by_provider}
\title{Graph Performance by Provider}
\usage{
graph_chart_by_provider()
}
\value{
plot element
}
\description{
Graphs each individual provider's performance over time as a ridgeline plot
Sorted by \% open more than 8 days, with \% open more than 24 hours included to side.
}
