% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_chart_stats.R
\name{graph_chart_stats}
\alias{graph_chart_stats}
\title{Graph Chart Stats}
\usage{
graph_chart_stats(remove_worst_performers = NULL)
}
\arguments{
\item{remove_worst_performers}{(integer) optionally removes the top n worst performers, base on most recent \% charts open more than 8 days.}
}
\value{
Plot item
}
\description{
Creates a chart of open chart rates for all times with gems showing specific current metrics.
}
