% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_new_data}
\alias{import_new_data}
\title{Loads Incidents for date range}
\usage{
import_new_data(start_date, end_date, source = "custom")
}
\arguments{
\item{start_date}{Date to start search}

\item{end_date}{Date to end search}

\item{source}{"custom" or "eso"}
}
\value{
Tibble of incidents
}
\description{
Includes incident information and crew information, along with timeframe chart was open for (interval)
}
